backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
python-dateutil==2.9.0.post0
pandas==2.2.3
aiofiles==24.1.0
httpx==0.28.1
